Parameter to show or not the mention to the user
# Addition of a parameter to show or not the mention to the user who made the pull request.

Note: Action is sensational, congratulations on the work done so far.

Motivation:
Use in personal projects, and to a certain extent, I work alone, but I like to keep the logs organized in releases, so with each successful pull request, I copy the changelog and paste it into a draft version. During this period, I see no reason to keep 'alerting' me, I don't mind, but I would like to be able to disable when necessary.